+++ src/twg2/treeLike/parameters/AbstractTreeTraverseParameters.java
@@ -4,9 +4,6 @@ import java.util.function.Function;
 import java.util.function.IntConsumer;
 import java.util.function.Predicate;
 
-import lombok.Getter;
-import lombok.Setter;
-import lombok.experimental.Accessors;
 import twg2.treeLike.TreeTraversalOrder;
 
 /**
@@ -16,19 +13,18 @@ import twg2.treeLike.TreeTraversalOrder;
  * @param <C> the type of children collection returned by the children getter
  * @param <F> the type of consumer function that visits each tree node
  */
-@Accessors(chain = true)
 public abstract class AbstractTreeTraverseParameters<T, S, C extends Iterable<? extends S>, F> implements TreeTraverseParameters<T, S, C, F> {
-	public @Getter @Setter T tree;
-	public @Getter @Setter boolean onlyVisitLeaves;
+	public T tree;
+	public boolean onlyVisitLeaves;
 	/** whether to skip the root node - default false */
-	public @Getter @Setter boolean skipRoot;
+	public boolean skipRoot;
 	/** whether to skip null root node - default false */
-	public @Getter @Setter boolean skipNullRoot;
-	public @Getter @Setter TreeTraversalOrder traversalOrder;
-	public @Getter @Setter Predicate<T> hasChildren;
-	public @Getter @Setter Function<T, C> childrenGetter;
-	public @Getter @Setter IntConsumer startSubtreeFunc;
-	public @Getter @Setter IntConsumer endSubtreeFunc;
+	public boolean skipNullRoot;
+	public TreeTraversalOrder traversalOrder;
+	public Predicate<T> hasChildren;
+	public Function<T, C> childrenGetter;
+	public IntConsumer startSubtreeFunc;
+	public IntConsumer endSubtreeFunc;
 
 
 	public AbstractTreeTraverseParameters(T tree, boolean onlyVisitLeaves, boolean skipNullRoot, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren,
@@ -43,59 +39,41 @@ public abstract class AbstractTreeTraverseParameters<T, S, C extends Iterable<? 
 		this.endSubtreeFunc = endSubtreeFunc;
 	}
 
-/*
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setTree(T tree) {
-		this.tree = tree;
-		return this;
-	}
 
+	@Override public T getTree() { return tree; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setOnlyVisitLeaves(boolean onlyVisitLeaves) {
-		this.onlyVisitLeaves = onlyVisitLeaves;
-		return this;
-	}
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setTree(T tree) { this.tree = tree; return this; }
 
+	@Override public boolean isOnlyVisitLeaves() { return onlyVisitLeaves; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setSkipNullRoot(boolean skipNullRoot) {
-		this.skipNullRoot = skipNullRoot;
-		return this;
-	}
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setOnlyVisitLeaves(boolean onlyVisitLeaves) { this.onlyVisitLeaves = onlyVisitLeaves; return this; }
 
+	@Override public boolean isSkipRoot() { return skipRoot; }
 
-	public AbstractTreeTraverseParameters<T, S, C, F> setTraversalOrder(TreeTraversalOrder traversalOrder) {
-		this.traversalOrder = traversalOrder;
-		return this;
-	}
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setSkipRoot(boolean skipRoot) { this.skipRoot = skipRoot; return this; }
 
+	@Override public boolean isSkipNullRoot() { return skipNullRoot; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setHasChildren(Predicate<T> hasChildren) {
-		this.hasChildren = hasChildren;
-		return this;
-	}
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setSkipNullRoot(boolean skipNullRoot) { this.skipNullRoot = skipNullRoot; return this; }
 
+	public TreeTraversalOrder getTraversalOrder() { return traversalOrder; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setChildrenGetter(Function<T, C> childrenGetter) {
-		this.childrenGetter = childrenGetter;
-		return this;
-	}
+	public AbstractTreeTraverseParameters<T, S, C, F> setTraversalOrder(TreeTraversalOrder traversalOrder) { this.traversalOrder = traversalOrder; return this; }
 
+	@Override public Predicate<T> getHasChildren() { return hasChildren; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setStartSubtreeFunc(IntConsumer startSubtreeFunc) {
-		this.startSubtreeFunc = startSubtreeFunc;
-		return this;
-	}
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setHasChildren(Predicate<T> hasChildren) { this.hasChildren = hasChildren; return this; }
 
+	@Override public Function<T, C> getChildrenGetter() { return childrenGetter; }
 
-	@Override
-	public AbstractTreeTraverseParameters<T, S, C, F> setEndSubtreeFunc(IntConsumer endSubtreeFunc) {
-		this.endSubtreeFunc = endSubtreeFunc;
-		return this;
-	}
-*/
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setChildrenGetter(Function<T, C> childrenGetter) { this.childrenGetter = childrenGetter; return this; }
+
+	@Override public IntConsumer getStartSubtreeFunc() { return startSubtreeFunc; }
+
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setStartSubtreeFunc(IntConsumer startSubtreeFunc) { this.startSubtreeFunc = startSubtreeFunc; return this; }
+
+	@Override public IntConsumer getEndSubtreeFunc() { return endSubtreeFunc; }
+
+	@Override public AbstractTreeTraverseParameters<T, S, C, F> setEndSubtreeFunc(IntConsumer endSubtreeFunc) { this.endSubtreeFunc = endSubtreeFunc; return this; }
+
 }