+++ src/twg2/treeLike/parameters/IndexedTreeTraverseParameters.java
@@ -4,9 +4,8 @@ import java.util.function.Function;
 import java.util.function.IntConsumer;
 import java.util.function.Predicate;
 
-import lombok.val;
 import twg2.collections.interfaces.ListReadOnly;
-import twg2.treeLike.IndexedSubtreeConsumer;
+import twg2.treeLike.IndexedTreeConsumer;
 import twg2.treeLike.TreeTraversalOrder;
 
 /**
@@ -14,9 +13,9 @@ import twg2.treeLike.TreeTraversalOrder;
  * @since 2015-8-31
  * @param <T> the type of data in each tree node
  */
-public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, ListReadOnly<? extends T>, IndexedSubtreeConsumer<T>> {
+public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParameters<T, T, ListReadOnly<? extends T>, IndexedTreeConsumer<T>> {
 	// package-private
-	IndexedSubtreeConsumer<T> consumer;
+	IndexedTreeConsumer<T> consumer;
 
 	@Override
 	public IndexedTreeTraverseParameters<T> setTree(T tree) { super.setTree(tree); return this; }
@@ -37,55 +36,55 @@ public class IndexedTreeTraverseParameters<T> extends AbstractTreeTraverseParame
 
 
 	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ListReadOnly<? extends T>> childrenGetter,
-			IndexedSubtreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
+			IndexedTreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
 		this(tree, onlyVisitLeaves, false, traversalOrder, hasChildren, childrenGetter, consumer, startSubtreeFunc, endSubtreeFunc);
 	}
 
 
 	public IndexedTreeTraverseParameters(T tree, boolean onlyVisitLeaves, boolean skipNullRoot, TreeTraversalOrder traversalOrder, Predicate<T> hasChildren, Function<T, ListReadOnly<? extends T>> childrenGetter,
-			IndexedSubtreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
+			IndexedTreeConsumer<T> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
 		super(tree, onlyVisitLeaves, skipNullRoot, traversalOrder, hasChildren, childrenGetter, startSubtreeFunc, endSubtreeFunc);
 		this.consumer = consumer;
 	}
 
 
 	@Override
-	public IndexedSubtreeConsumer<T> getConsumer() {
+	public IndexedTreeConsumer<T> getConsumer() {
 		return this.consumer;
 	}
 
 
 	@Override
-	public IndexedTreeTraverseParameters<T> setConsumer(IndexedSubtreeConsumer<T> consumer) {
+	public IndexedTreeTraverseParameters<T> setConsumer(IndexedTreeConsumer<T> consumer) {
 		return setConsumerIndexed(consumer);
 	}
 
 
-	public IndexedSubtreeConsumer<T> getConsumerIndexed() {
+	public IndexedTreeConsumer<T> getConsumerIndexed() {
 		return this.consumer;
 	}
 
 
-	public IndexedTreeTraverseParameters<T> setConsumerIndexed(IndexedSubtreeConsumer<T> consumer) {
+	public IndexedTreeTraverseParameters<T> setConsumerIndexed(IndexedTreeConsumer<T> consumer) {
 		this.consumer = consumer;
 		return this;
 	}
 
 
 	public static <_T> IndexedTreeTraverseParameters<_T> of(_T tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
-		val params = new IndexedTreeTraverseParameters<_T>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
+		var params = new IndexedTreeTraverseParameters<_T>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
 	public static <_T> IndexedTreeTraverseParameters<_T> leafNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
-		val params = new IndexedTreeTraverseParameters<_T>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
+		var params = new IndexedTreeTraverseParameters<_T>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
 	public static <_T> IndexedTreeTraverseParameters<_T> allNodes(_T tree, TreeTraversalOrder traversalOrder, Predicate<_T> hasChildren, Function<_T, ListReadOnly<? extends _T>> childrenGetter) {
-		val params = new IndexedTreeTraverseParameters<_T>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
+		var params = new IndexedTreeTraverseParameters<_T>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}