+++ src/twg2/treeLike/parameters/KeyTreeTraverseParameters.java
@@ -6,8 +6,7 @@ import java.util.function.Function;
 import java.util.function.IntConsumer;
 import java.util.function.Predicate;
 
-import lombok.val;
-import twg2.treeLike.SubtreeConsumer;
+import twg2.treeLike.TreeConsumer;
 import twg2.treeLike.TreeTraversalOrder;
 
 /**
@@ -15,9 +14,9 @@ import twg2.treeLike.TreeTraversalOrder;
  * @since 2015-8-31
  * @see AbstractTreeTraverseParameters
  */
-public class KeyTreeTraverseParameters<K, V> extends AbstractTreeTraverseParameters<V, Map.Entry<K, V>, Iterable<? extends Map.Entry<K, V>>, SubtreeConsumer<Map.Entry<K, V>>> {
+public class KeyTreeTraverseParameters<K, V> extends AbstractTreeTraverseParameters<V, Map.Entry<K, V>, Iterable<? extends Map.Entry<K, V>>, TreeConsumer<Map.Entry<K, V>>> {
 	// package-private
-	SubtreeConsumer<Entry<K, V>> consumer;
+	TreeConsumer<Entry<K, V>> consumer;
 
 	@Override
 	public KeyTreeTraverseParameters<K, V> setTree(V tree) { super.setTree(tree); return this; }
@@ -39,20 +38,20 @@ public class KeyTreeTraverseParameters<K, V> extends AbstractTreeTraverseParamet
 
 	@SuppressWarnings("unchecked")
 	public KeyTreeTraverseParameters(V tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<V> hasChildren, Function<V, ? extends Iterable<? extends Map.Entry<? extends K, ? extends V>>> childrenGetter,
-			SubtreeConsumer<Map.Entry<K, V>> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
+			TreeConsumer<Map.Entry<K, V>> consumer, IntConsumer startSubtreeFunc, IntConsumer endSubtreeFunc) {
 		super(tree, onlyVisitLeaves, false, traversalOrder, hasChildren, (Function<V, Iterable<? extends Map.Entry<K, V>>>) childrenGetter, startSubtreeFunc, endSubtreeFunc);
 		this.consumer = consumer;
 	}
 
 
 	@Override
-	public SubtreeConsumer<Entry<K, V>> getConsumer() {
+	public TreeConsumer<Entry<K, V>> getConsumer() {
 		return this.consumer;
 	}
 
 
 	@Override
-	public KeyTreeTraverseParameters<K, V> setConsumer(SubtreeConsumer<Entry<K, V>> consumer) {
+	public KeyTreeTraverseParameters<K, V> setConsumer(TreeConsumer<Entry<K, V>> consumer) {
 		this.consumer = consumer;
 		return this;
 	}
@@ -60,19 +59,21 @@ public class KeyTreeTraverseParameters<K, V> extends AbstractTreeTraverseParamet
 
 	public static <F, G> KeyTreeTraverseParameters<F, G> of(G tree, boolean onlyVisitLeaves, TreeTraversalOrder traversalOrder, Predicate<G> hasChildren,
 			Function<G, ? extends Iterable<? extends Map.Entry<F, G>>> childrenGetter) {
-		val params = new KeyTreeTraverseParameters<F, G>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
+		var params = new KeyTreeTraverseParameters<F, G>(tree, onlyVisitLeaves, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
-	public static <F, G> KeyTreeTraverseParameters<F, G> leafNodes(G tree, TreeTraversalOrder traversalOrder, Predicate<G> hasChildren, Function<G, ? extends Iterable<? extends Map.Entry<? extends F, ? extends G>>> childrenGetter) {
-		val params = new KeyTreeTraverseParameters<F, G>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
+	public static <F, G> KeyTreeTraverseParameters<F, G> leafNodes(G tree, TreeTraversalOrder traversalOrder, Predicate<G> hasChildren,
+			Function<G, ? extends Iterable<? extends Map.Entry<? extends F, ? extends G>>> childrenGetter) {
+		var params = new KeyTreeTraverseParameters<F, G>(tree, true, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}
 
 
-	public static <F, G> KeyTreeTraverseParameters<F, G> allNodes(G tree, TreeTraversalOrder traversalOrder, Predicate<G> hasChildren, Function<G, ? extends Iterable<? extends Map.Entry<? extends F, ? extends G>>> childrenGetter) {
-		val params = new KeyTreeTraverseParameters<F, G>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
+	public static <F, G> KeyTreeTraverseParameters<F, G> allNodes(G tree, TreeTraversalOrder traversalOrder, Predicate<G> hasChildren,
+			Function<G, ? extends Iterable<? extends Map.Entry<? extends F, ? extends G>>> childrenGetter) {
+		var params = new KeyTreeTraverseParameters<F, G>(tree, false, traversalOrder, hasChildren, childrenGetter, null, null, null);
 		return params;
 	}